project: Pass `ServiceConfig` to `RequiredExternalTools`
This modifies the `RequiredExternalTools` method on both the
`FrameworkService` and `ServiceTarget` interfaces to accept the
`ServiceConfig` of the service in question we are fetching tools for.

This allows the set of tools a to be conditional based on the service
configuration, which we will exploit later when supporting docker
remote builds via ACR so we can say the docker CLI is not a required
tool.