tools: Remove `BicepCli` interface (#4116)
When we started to depend on external tools, we used a pattern of
defining an interface to represent the capabilities of the tool which
the rest of the system dependend on.  The thought was having this
interface indirection would help us with unit-testing, because we
could mock different parts of the tool directly instead of depending
on mocks via the external command runner.

In practice, we never did this, when we test things, we just mock at
the `exec.CommandRunner` layer or use something like the recorder to
capture all the external tool interactions and replay them later.

Since we don't have multiple implementations of the interface, let's
just remove it and write things in the more direct style as we haven
been doing more recently where we don't have the intermediate
interface.

With this change, `bicep.BicepCli` (an interface) is now replaced with
`*bicep.Cli` which is logically what was returned in the past anyway,
but behind the guise of the interface).