bicep: Fix panic during deployment progress rendering
We had a panic in our display code when rendering data if you were
using a the microsoft graph provider in bicep (a new extensibility
feature) because some of the properties of the ARM operation object
were not present.

This change adds additional guards to our rendering logic to ignore
operations that don't have a complete set of data.

Fixes #4169