chore: Remove "lowercase c" pattern from the codebase
Early on in the project, we used to use this pattern where we'd
prepend a lowercase c in front of `const` values (and var values which
were logical constants but couldn't be marked as `const` because their
values were not compile time constants). We did this to ensure the
symbol was not exported but to still have something
`ThatLookedLikeItWouldBeExported` with the idea that this would make
it more obvious in code this was a constant value.

However, idiomatic go code doesn't do this, so we stopped doing it
ourselves on new code. This change does a pass over the old pattern
and removes it from the codebase. In most cases I just dropped the
leading c and lowercased the first letter (so the symbol is not
exported) but in a few cases where the symbol was only used in one or
two places I inlined it or rewrote code slightly if I felt it would
improve overall clarity.