azcli: Remove some dead code
When we started the project, we used the `azcli` package to run an
instance of `az` to do most of our management plane operations. Over
time we moved away from this (mostly rewrting it implementation on top
of `azure-sdk-for-go` packages) and as we did this we moved a bunch of
the logic out of the `azcli` package and into other packages.

As we did that, we ended up leaving some dead code behind that wasn't
obviously dead (either becaue it was exported but not consumed or
because our DI pattern makes it non obvious when an injected value is
unused because it is written to via an exported constructor).

This change cleans up this dead code, simplifying the codebase.