Remove `azdsdk.ClientOptionsBuilder` and its factory
Instead of using the builder pattern, it is often cleaer to just
construct the `*azcore.ClientOptions` object directly in-line. In most
cases, folks end up just injecting and `*arm.ClientOptions` or
`*azcore.ClientOptions` anyway, and seeing the configuration of them
in `container.go` instead of the policy being hidden behind some
builder is also nice.