project: Add support for Java functions (#4208)
This change adds support for deploying Java applications to `function` host.

When packaging, the default `target/azure-functions/$functionAppName` directory will be used
for zip deployment, where `$functionAppName` is sourced from the maven project. If there is a 
single directory `target/azure-functions/<some directory>`, that directory is used instead.

Otherwise, the user is able to specify `dist` to the location of where their custom staging directory
is.  
 
Addresses #4155