maven: Use `exec.ErrNotFound` in `errors.Is`
`exec.LookPath` returns `exec.ErrNotFound` when it fails to find a
binary, but we were testing against `os.ErrNotExist`. In versions
previous to go 1.23 this worked out, but in go 1.23 it no longer does
(and a handful of the unit tests around this code started to fail), so
we need to start testing against the correct error.