exec: Don't depend on internals of `os.Process`
To assign the process to the job object after we create it, we need
the `windows.Handle` that backs the process we started with `os/exec`.
This was not exported and so we used some gnarly casting to cast the
`os.Process` to a structure we authored that had the same layout of
`os.Process` so we could grab the handle property.

This is unsafe if the layout of the `os.Process` changes, which it did
in Go 1.23.

Move to code that doesn't depend on the internals of `os.Process` by
using the exported Pid to call `windows.OpenProcess` to get a handle and
then pass that along to `windows.AssignProcessToJobObject`.