recording: Improvements to `Content-Length` handeling (#4231)
When you call the `blockblob.GetProperties`, the azure-sdk-for-go
makes a HEAD request to fetch information about the target blob, and
uses the result to hydrate the properties object that is returned.
The `ContentLength` property (which tells you how big the blob is) is
set based on the `Content-Length` header in the response. The type of
`ContentLength` is `*int` and if the `Content-Length` header is not
set in the response, then the value will remain `nil`.

As part of the new ACR remote build feature, to stream logs from the
remote build to the console we use `blockblob.GetProperties` to fetch
the length of the remote build log file and then do a range request to
fetch the content that we have not yet processed. When I wrote an end
to end test and tried to record it, this code ended up crashing with a
nil reference because the `ContentLength` property was `nil`.

This ended up being due to logic in the recording proxy which uses
chuncked encoding when returning the response from the proxy server to
the client. Per the HTTP spec, when using chunked encoding, the
response should not have a Content-Length header, and the indeed the
go standard library removes the header before writing the response.
Talking with Wei, we think this code to use chunked response was added
early on in the development of the proxy to work around some content
length mismatches he was facing and may not be needed long term.

For now, I've updated the logic to not used chunked encoding for HEAD
requests. Since HEAD requests to have a body there's no real value in
using chunked responses and doing this ensures the Content-Length
header is returned in the response.

While in the area, I also updated the middleware that we have to
expand gzip'd responses (so that the recording contains the expanded
content) to update the Content-Length property when present to be the
length of the expanded content instead of just dropping it. While
investigating this issue I first thought this was where the problem
was, but it ended up being the chunked encoding, the HEAD response
returned by Blob Storage is not gzip encoded.