As part of merging azd operation and hooks, this PR allows users to define hooks outside of azure.yaml as individual files with the name azd.hooks.[yaml | yml].
The file can be placed in the /infra folder for project level hooks or inside service's folder for service level hooks.

This will allow generators like Aspire to include hooks as part of the infrastructure code (w/o touching azure.yaml).  Then, when running azd infra synth, the hooks are part of the infrastructure and users can delete the infra folder to go back to full Aspire in-memory generation w/o touching azure.yaml.

After this PR, I am planning to move the azd operations to become a valid built-in hook and delete the alpha feature for azd operations and the entire feature in favor of just using the hooks strategy