recording: Use `WithHostMapping` in `TestBlobClientGetProperties`
When developing the test I unintentionally set
`recorder.ModeRecordOnly` causing us to recrate the recording per
run. Turning that off exposed an issue where since the `httptest`
package picks a different port each run it meant that the recording
would not match because the port was different, causing us to fail to
playback the test.

To address this, I've introduced a new option to the recorder,
`WithHostMapping` that allows you to map one host to another when
persisting recordings and then used it in the test.