relax ghrepo validation (#4237)
* pin cspell version due to issue on last release

* Relax detection of github host names during azd pipeline config.
Folks can define an alias for a hostname to control the usage of different ssh keys. When this happens, the hostname will not be github.com, but it might be still pointing to github.
This change relaxes the validation to allow any valid hostname. The command would fail later if the host is not accessible.

* Relax detection of github host names during azd pipeline config.
Folks can define an alias for a hostname to control the usage of different ssh keys. When this happens, the hostname will not be github.com, but it might be still pointing to github.
This change relaxes the validation to allow any valid hostname. The command would fail later if the host is not accessible.