WIP: aspire: Use bicep for container app deployment
The plan for Aspire 9 is to support customization of the container app
for services by leveraging the CDK to generate the bicep IaC for a
given container app.

To prepare for that new world, we'd like `azd infra synth` to start
writing bicep and to modify our deployment strategy for container apps
to be based on ARM deployments instead of directly PUTing a copy of
the resource to the RP.

This change starts to work through that.