WIP: aspire: Support bicep based container app deployments
This adds rudimentary support for deploying the `.bicep` version of
the IaC for a container app as generated by `azd infra synth`.

In the case where both a `.yaml` and `.bicep` version of the IaC for a
container app exist, the container `.yaml` version is prefered.

When neither the `.yaml` or `.bicep` version of the IaC exist on disk,
`azd` will continue to generate and deploy the `.yaml` flavored
version of the container app IaC.