chore: Update comments in `launch.json` (#4303)
I was talking with Marina today about my workflow for debugging
changes in `azd`. I use the "Debug azd cli" launch configuration for
most of my debugging and end up modifying `launch.json` to set the
`cwd` and other parameters for the profile for the specific end to end
I want to debug.  There are some `git` tricks that you can play to
ignore modifications to this file (so you don't have to worry about
checking them in).  I had documented them in one place but not the
other (since I typically run VS Code from within the `cli/azd` sub
folder, to have a smaller workspace) and didn't document how you undo
this behavior.

This patch updates the comments to show how to turn this off, and adds
them to our root `launch.json` as well.