WIP: provisioning: Consider `resourceGroup` from `azure.yaml`
We allow specifying the resource group which is used for discovery of
resources (instead of probing for a resource group taged with
`azd-env-name` inside `azure.yaml`. When doing a resource group scoped
deployment, we should consider this value if set and use it instead of
prompting the user to select a resource group or create a new one.

Fixes #4287