remove bicep `main.parameters.json` caching (#4363)
Remove in-memory `main.parameters.json` cache.

By removing this in-memory cache, we incur an extra load from disk, but without any noticeable behavioral changes.

Background: In-memory caching for `main.parameters.json` was added initially to avoid a re-prompt when the user declines the prompt "Save the value in the environment for future use?" (https://github.com/Azure/azure-dev/pull/3155). This was added as a way to communicate within `azd` that "this value is configured for the current azd process lifetime". This was superseded by the changes in https://github.com/Azure/azure-dev/pull/3479 where we always saved the values to `config.json` and removed the prompt. 

Fixes #3973, #4310