Updates `azure.yaml` schema for deployment stacks configuration (#4358)
Updates the alpha schema for azure.yaml to include the optional deployment stacks configuration.

The deploymentStacks property will light up when provider: bicep. When specifying the deploymentStacks property the user must configure at least one of the properties otherwise the whole property can be omitted and use the internal defaults.