Require resourceType for resource prompt
Within a template - hence only checking during parameter processing - prompting for *any* resource is practically useless. You have to know the type to use as intended e.g., an Azure Monitor resource. To even use an existing resource in `reference()` or `resourceId()` functions you need to know the type, as do you in an `existing` resource template in Bicep.