ci: Update bicep in lint-bicep to v0.31.92
The current version that is installed on the runners (0.31.34) has
some bugs around linting and causes lint warnings during our
legs. These errors are not present on the latest Bicep CLI.

Force an upgrade of the binary that is installed as part of the test
runner to latest version. We should be able to remove this step once
new image is built with an upgraded bicep (the `README.md` at
https://github.com/actions/runner-images/tree/main/images/ubuntu has
information about what is on the runner.)