docs: Add documentation for `azd infra synth`
We added `azd infra synth` as an alpha command which can be used for
Aspire and composibility based project to write the generated
infrastructure that is normally managed by `azd` to disk so it can be
tweaked by an end user.

This command was not showing up in our documentation, since we hide
the entire `infra` tree, since previous to `infra synth` the only
other two commands under the `infra` command where `create` and
`delete` and we had deprecated these commands (in favor of `provision`
and `down`).

This change adjusts things such that `infra synth` correctly shows up
in the help text from Cobra (so `docgen` will include it in the
generated documentation we ship to learn.microsoft.com) and correctly
marks it as "(Alpha)" instead of "(Beta)".

Fixes #4590