publish: Stop publishing "All In One" container (#4616)
As part of our release process, we build and push a container image
that contains `azd` plus a large collection of language runtimes and
other tools.

In practice, the resulting image is giant and anyone who wants to use
`azd` inside of a container is better served by building their own and
including `azd`. We also have not been updating this container over
time (and a while back we removed it from the list of artifacts we
list per PR).

To make our lives simpler, let's just stop building and publishing
this container image. If folks were depedent on it and need updates,
let's move them over to using one of our supported install methods to
get them to add `azd` to their container image as the build it instead
of using ours.