refactor: Move remaining code in `pkg/tools/azcli/` to `pkg/azapi`
When we started `azd` instead of making REST calls against the Azure
using the SDKs, we had an `AzCli` type which shelled out to `az` to
preform work. Over time we refactored this code from using the `exec`
package to run `az` to use the proper go SDKs from azure-sdk-for-go
and as we did this we ended up pulling off related operations into
their own services and packages.

Over time we got close to completely removing the `AzCli` type, but a
few operations still lingered.

To finish the job while minimizing the destruction to the rest of the
codebase, I've taken the `AzCli` type in `pkg/tools/azcli/`, moved all
the code into `pkg/azapi` and then renamed the type from `AzCli` to
`AzureClient` (following the azure-sdk-for-go convention of `Client`
suffixes). While doing so, I removed the intermediate wrapper
interface and now other components just take a `*azapi.AzureClient`
instead of a `azcli.AzCli`.

Fixes #854