rpc: Support handlers of any arity
Previously, we had of hand written adapters that allowed exposing go
functions over our JSON-RPC boundary.  Changes to the number of
arguments or return values would require changes to the adapter used
(or writing a new adapter by copy/pasting an existing one).

Now, we have a single `NewHandler` which allows exposing arbitrary go
functions over the RPC boundary (as long as the follow some rules,
like taking a `context.Context` as the first paremeter, and returning
a value and `error` or just an `error`.

As part of this change, the `IObserver` type is now called `Observer`
and it is passed by reference instead of value in signatures.