bicep: Support `@sealed()` decorator on user defined types (#4684)
When the `@sealed()` decorator is applied on user defined types, it
causes an `additionalProperties: false` property to be added to the
type defintition in the generated arm deployment. Our unmarshalling
logic did not understand this pattern and so it would cause an error
when trying to unmarshall this, since it expected if
`additionalProperties` was set, it would be to set to an object that
described information about the additional properties.

I've updated the marshalling code to support `false` as a valid value
for the additional properties node when parsing an ARM template. Since
Go doesn't have great support for modeling descrimited unions, the new
`ArmTemplateParameterAdditionalPropertiesValue` type is used to
provide the logcal descrimited union of `false` or a object like
value.

Fixes #4659