tests: Remove Terraform Remote State integration test
When we were bringing up the `azd` support for Terraform, we added two
end to end tests which deployed an application using Terraform.  One
stored the state file localy and the other used the terraform remote
state backend to store the state in blob storage.

When we migrated to the TME environment, this broke, because we need
to use OIDC to authenticate to the storage account, but the remote
state plugin isn't able to do this natively, and the support to use
the `az` CLI fails in in this mode because it is only supported for
user accounts (and the tests run under a service principal).

In practice, `azd` isn't doing anything special here - the test uses
the `az` CLI to create a storage account and then sets some
environment variables to configure the terraform provider to use it
for state management.  So, instead of fighting with the infrastructure
to support this tests, let's just get rid of it.  We already have end
to end coverage of Terraform itself with the local state test, and the
way `azd` interacts with terraform is the same between remote and
local state storage, so we aren't getting any additional coverage from
the test.

Closes #4564