chore: `init` to depend on `add.Configure` (#4698)
Reshares UI prompting logic between `init` and `add`.

`init` can be thought of as a series of `add` calls, though this is not reflected in practice due to difference in UX design considerations.

This is a pure refactor and does not introduce any functional changes. The UI prompting is covered by existing tests.