compose: show resources (#4914)
Add ability to `azd show <name>` for all existing supported resources.

With this change, we add the ability to display fully-evaluated connection variables for resources without provisioning a `host` resource.

To do this, we create a generic mapping between ARM and the app representation. This mapping, which is comprised of simple templating-like expressions, are evaluated against live Azure in the `show` command. 

Future ideas:

1. The underlying logic for `azd show <full resource ID>` could be used to display, or as a base layer, for obtaining connection information to an existing resource. The schema does not track managed identity vs key-based approaches, but I do see a somewhat straightforward pathway there.

2. The schema mapping could be used to *emit* Bicep representation during generation. This may end up unifying the story between generation and evaluation for roundtripping.

3. We expect to use parts of the parsing logic in to support a similar expression-model at the app layer.

Closes #4727
Contributes to #4577, #4747, #4887