compose: ai endpoint (#4966)
Fix `AZURE_OPENAI_ENDPOINT` not showing in add preview.

In #4914, we shifted the preview to leverage the newly added resource metadata. Azure OpenAI models had a resource specific show implementation built that wasn't immediately refactored in the change.

This change shifts the Azure OpenAI implementation and fixes the add preview.

Contributes to #4933