Pass console with to extensions when invoked (#5132)
When a command is invoked through a sub process/shell, the called extension is unable to determine the correct console width using the consolesize.GetConsoleSize() that is used within azd.

To work around this issue, the currently detected console width will be set in a COLUMNS environment variable and then read by the extensions if unable to detect the size using standard libraries.

The COLUMNS environment variable is a semi-standard used in some Mac/unix based terminals. If a console width is still unable to be detected we fallback to a default of 120.