compose - AI Foundry 1RP changes (#5157)
Composability changes to support AI Foundry's new 1RP - basic setup (no capability host).

During the first provision, azd will check quota for all Foundry models added to the project and will prompt for a single location. Those models get deployed under a single AI Services account in that location.

With 1RP, there is still a concept of a project, but it's now a child resource of the AI Services account.

---------

Co-authored-by: Victor Vazquez <vhvb1989@gmail.com>