fix: `infra gen` when dotnet project is present
This change adds missing logic to check if a dotnet project is an Aspire
project before attempting to run the `dotnet run` command.

Before thiBs change, when users attempted to run `infra gen` with a dotnet project
present; the `dotnet run` command would end up running the application itself,
which results in unexpected behavior such as azd hanging while waiting for the application
to return.