vscode: fix handling of azure.yaml directory movements
Problem: user is currently trapped in repeated error dialog prompts when a directory containing
azure.yaml is deleted, or renamed.

This fix addresses the problem by:

1. Subscribing for azure.yaml configuration directory changes as they occur.
2. Change refresh button behavior to properly rediscover all azd applications in current workspace(s).
This allows user to perform a manual refresh in case of any missed updates.

Underneath-the-covers, the AzureDevApplicationProvider is made to be a stateless data provider
to support the scenario as described in 2. above.

Fixes: #5690