Fixes lifetime issues with container helper from extension container service (#6028)
Resolves #6019 - This change updates how the azd environment is resolved from within the container helper. Instead of relying on the environment field of the ContainerHelper we refactor to take a dependency of the environment manager and then fetch the environment before usage.

The original issue was caused by the container helper being loaded within a singleton container service for the extension framework. Since this service is long lived throughout the full azd invocation this was causing issues in workflow based scenarios where a stale or cached azd environment was being referenced instead of the correct scoped instance.

This ensures that the azd environment we are working against is always an up to date version that includes all variables and configuration.